Правки нейминга и обертки парсера
- Обертка парсера теперь бросает специальное исключение, которое ловится при её использовании
- Изменена генерация автомата для языка суффиксов
